pyproject.toml
[project]
name = "claude-api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.62.0",
    "python-dotenv>=1.1.1",
]
